BlitzSSH/requirements.txt
paramiko 
python-telegram-bot 
rich 
pystyle